services/eagle-eye/pyproject.toml
[tool.poetry]
name = "clestiq-shield-eagle-eye"
version = "0.1.0"
description = "Authentication and Application Management Service for Clestiq Shield"
authors = ["Clestiq Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
email-validator = "^2.1.0"
argon2-cffi = "^21.3.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
httpx = "^0.26.0"
pydantic-settings = "^2.1.0"
structlog = "^24.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
